Take resource group name from config.json
I'd previously thought that to simplify the config file
we would generate the resource group name from the name
of the binderhub instance, but when deployment happens
from blue button the resource group name is passed in
to the deploy container, so to standardise we should use
resource group name from the config file.

This commit removes a couple of lines that should have
been taken out when the move to config file resource
group naming happened.